// console-master/partials/bis/dashboard/src/view/types.ts
export enum BisDashboardWidgetType {
  AlertSummary = 'BIS-alert-summary',
  IdentityRiskAlerts = 'BIS-identity-risk-alerts',
  GeozoneRiskAlerts = 'BIS-geozone-risk-alerts',
  CombinedRiskAlerts = 'BIS-combined-risk-alerts',
  RealTimeAlerts = 'BIS-real-time-alerts',
  TopActions = 'BIS-top-actions',
  TotalEvents = 'BIS-total-events',
  TotalUsers = 'BIS-total-users',
  Map = 'BIS-map',
}